refact:signup 기능 통합
