+++ frontend/src/apis/request/requestAuth.ts
@@ -1,12 +1,16 @@
 import { axios } from '../axios';
 import { API_PATH } from '@/constants/path';
 
-const requestLoginByEmail = (userData: RequestLoginByEmailType) => {
-  return axios.post<ResponseLogin>(API_PATH.AUTH.LOGIN_BY_EMAIL, userData);
+const requestLogin = (
+  userData: RequestLoginByEmailType | RequestLoginByPhoneType,
+  type: string,
+) => {
+  return axios.post<ResponseLogin>(
+    type === 'email'
+      ? API_PATH.AUTH.LOGIN_BY_EMAIL
+      : API_PATH.AUTH.LOGIN_BY_PHONE,
+    userData,
+  );
 };
 
-const requestLoginByPhone = (userData: RequestLoginByPhoneType) => {
-  return axios.post<ResponseLogin>(API_PATH.AUTH.LOGIN_BY_PHONE, userData);
-};
-
-export { requestLoginByEmail, requestLoginByPhone };
+export { requestLogin };
