+++ frontend/src/mock/handlers.ts
@@ -2,7 +2,6 @@ import { HttpResponse, http } from 'msw';
 
 export const handlers = [
   http.post('http://localhost:3000/login/*', ({ request }) => {
-    console.log(request);
     return HttpResponse.json(
       {
         userId: 1,
@@ -11,7 +10,6 @@ export const handlers = [
     );
   }),
   http.post('http://localhost:3000/signup/*', ({ request }) => {
-    console.log(request);
     return new HttpResponse(null, { status: 201 });
   }),
 ];
