+++ frontend/src/routes/Router.tsx
@@ -4,22 +4,14 @@ import CreateBandPage from '@/pages/CreateBandPage/CreateBandPage';
 import HomePage from '@/pages/HomePage/HomePage';
 import LoginPage from '@/pages/LoginPage/LoginPage';
 import { BROWSER_PATH } from '@/constants/path';
-import PhoneSignUpPage from '@/pages/SignUpPage/PhoneSignUpPage/PhoneSignUpPage';
-import EmailSignUpPage from '@/pages/SignUpPage/EmailSignUpPage/EmailSignUpPage';
+import SignUpPage from '@/pages/SignUpPage/SignUpPage';
 
 export default function Router() {
   return (
     <Routes>
       <Route path={BROWSER_PATH.HOME} element={<HomePage />} />
-      <Route path={BROWSER_PATH.LOGIN_BY_EMAIL} element={<LoginPage />} />
-      <Route
-        path={BROWSER_PATH.SIGN_UP_BY_PHONE}
-        element={<PhoneSignUpPage />}
-      />
-      <Route
-        path={BROWSER_PATH.SIGN_UP_BY_EMAIL}
-        element={<EmailSignUpPage />}
-      />
+      <Route path={BROWSER_PATH.LOGIN} element={<LoginPage />} />
+      <Route path={BROWSER_PATH.SIGN_UP} element={<SignUpPage />} />
       <Route path="/band:bandId" element={<BandDetailPage />} />
       <Route path={BROWSER_PATH.CREATE_BAND} element={<CreateBandPage />} />
     </Routes>
